+++ db/migrate/20140506124830_add_hide_values_on_pdfs_to_projects.rb
@@ -1,0 +1,5 @@
+class AddHideValuesOnPdfsToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :hide_values_on_pdfs, :boolean, null: false, default: false
+  end
+end