+++ lib/valuables/multiple_response.rb
@@ -5,11 +5,21 @@ module Valuables
   class MultipleResponse < Default
 
     def name
-      @object.variable.shared_options_select_values(@object.responses.pluck(:value)).collect{|option| option[:value] + ": " + option[:name]}
+      response_options.collect{|option| option[:value] + ": " + option[:name]}
     end
 
     def raw
-      @object.variable.shared_options_select_values(@object.responses.pluck(:value)).collect{|option| option[:value]}
+      response_options.collect{|option| option[:value]}
+    end
+
+    def display_name
+      response_options.collect{|option| option[:name]}
+    end
+
+    private
+
+    def response_options
+      @object.variable.shared_options_select_values(@object.responses.pluck(:value))
     end
 
   end