+++ lib/valuables/single_response.rb
@@ -12,6 +12,10 @@ module Valuables
       begin Integer(@object.response) end rescue @object.response
     end
 
+    def display_name
+      hash_display_name
+    end
+
   end
 
 end